pi/requirements.txt
pyrow @ git+https://github.com/droogmic/Py3Row.git
paho-mqtt>=2.0
